rm executor package
