+++ dlrover/python/rl/master/job_manager.py
@@ -11,15 +11,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from rl.master.executor import Executor
+from rl.master.graph import RLExecutionGraph
+from rl.master.scheduler import GroupOrderedScheduler, SimpleScheduler
+
 from dlrover.python.common.log import default_logger as logger
 from dlrover.python.rl.common.enums import SchedulingStrategyType
 from dlrover.python.rl.common.job_context import get_job_context
-from dlrover.python.rl.master.execution.executor import Executor
-from dlrover.python.rl.master.execution.graph import RLExecutionGraph
-from dlrover.python.rl.master.execution.scheduler import (
-    GroupOrderedScheduler,
-    SimpleScheduler,
-)
 
 
 class JobManager(object):
