+++ dlrover/python/rl/master/scheduler.py
@@ -19,6 +19,11 @@ import ray
 from ray.actor import ActorHandle
 from ray.exceptions import GetTimeoutError
 from ray.util.scheduling_strategies import PlacementGroupSchedulingStrategy
+from rl.master.graph import (
+    PlacementGroupAllocation,
+    RLExecutionGraph,
+    RLExecutionVertex,
+)
 
 from dlrover.python.common.log import default_logger as logger
 from dlrover.python.common.resource import Resource
@@ -26,11 +31,6 @@ from dlrover.python.rl.common.constant import RLMasterConstant
 from dlrover.python.rl.common.enums import RLRoleType, WorkloadGroupType
 from dlrover.python.rl.common.exception import ResourceError
 from dlrover.python.rl.common.job_context import get_job_context
-from dlrover.python.rl.master.execution.graph import (
-    PlacementGroupAllocation,
-    RLExecutionGraph,
-    RLExecutionVertex,
-)
 
 _job_ctx = get_job_context()
