+++ dlrover/python/rl/tests/master/test_executor.py
@@ -1,4 +1,4 @@
-# Copyright 2025 The EasyDL Authors. All rights reserved.
+# Copyright 2025 The DLRover Authors. All rights reserved.
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -10,3 +10,17 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from unittest.mock import MagicMock
+
+from rl.master.executor import Executor
+from rl.master.graph import RLExecutionGraph
+
+from dlrover.python.rl.tests.master.base import BaseMasterTest
+
+
+class ExecutorTest(BaseMasterTest):
+    def test_execute(self):
+        graph = RLExecutionGraph(self._job_context.rl_context)
+        executor = Executor(graph, None)
+        executor.create_workloads = MagicMock(return_value=None)
+        executor.execute()
