+++ dlrover/python/rl/tests/master/test_scheduler.py
@@ -14,16 +14,13 @@ import unittest
 from unittest.mock import MagicMock
 
 import ray
+from rl.master.graph import RLExecutionGraph
+from rl.master.scheduler import GroupOrderedScheduler, SimpleScheduler
 
 from dlrover.python.rl.common.args import parse_job_args
 from dlrover.python.rl.common.config import JobConfig
 from dlrover.python.rl.common.job_context import get_job_context
 from dlrover.python.rl.common.rl_context import RLContext
-from dlrover.python.rl.master.execution.graph import RLExecutionGraph
-from dlrover.python.rl.master.execution.scheduler import (
-    GroupOrderedScheduler,
-    SimpleScheduler,
-)
 from dlrover.python.rl.tests.master.base import BaseMasterTest
 from dlrover.python.rl.tests.test_data import TestData
